Integrate rag related operators into one.
.
